added edits

+++ src/models/entry-model.mjs
@@ -14,7 +14,6 @@ const listAllEntries = async () => {
   }
 };
 
-
 const findEntryById = async (id) => {
   try {
     const [rows] = await promisePool.query(
@@ -45,18 +44,25 @@ const addEntry = async (entry) => {
   }
 };
 
-const updateEntryById = async (entry) => {
-  const {entry_id, entry_date, mood, weight, sleep_hours, notes} = entry;
+const updateEntryById = async (entryId, userId, entryData) => {
   try {
-    const sql =
-      'UPDATE DiaryEntries SET entry_date=?, mood=?, weight=?, sleep_hours=?, notes=? WHERE entry_id=?';
-    const params = [entry_date, mood, weight, sleep_hours, notes, entry_id];
+    const params = [entryData, entryId, userId];
+    // format() function is used to include only the fields that exists
+    // in the entryData object to the SQL query
+    const sql = promisePool.format(
+      `UPDATE DiaryEntries SET ?
+       WHERE entry_id=? AND user_id=?`,
+      params
+    );
     const [result] = await promisePool.query(sql, params);
+    // console.log(result);
     if (result.affectedRows === 0) {
-      return {error: 404, message: 'entry not found'};
+      return {error: 404, message: 'Entry not found'};
     }
-    return {message: 'entry data updated', entry_id};
+    return {message: 'Entry data updated', entry_id: entryId};
   } catch (error) {
+    // fix error handling
+    // now duplicate entry error is generic 500 error, should be fixed to 400 ?
     console.error('updateEntryById', error);
     return {error: 500, message: 'db error'};
   }
@@ -67,7 +73,14 @@ const updateMedicationById = async (entry) => {
   try {
     const sql =
       'UPDATE Medications SET name=?, dosage=?, frequency=?, start_date=?, end_date=? WHERE medication_id=?';
-    const params = [name, dosage, frequency, start_date, end_date, medication_id];
+    const params = [
+      name,
+      dosage,
+      frequency,
+      start_date,
+      end_date,
+      medication_id,
+    ];
     const [result] = await promisePool.query(sql, params);
     // console.log(result);
     if (result.affectedRows === 0) {
@@ -83,11 +96,27 @@ const updateMedicationById = async (entry) => {
 };
 
 const updateNutritionById = async (entry) => {
-  const {nutrition_id, entry_date, calories_consumed, protein_grams, carbohydrates_grams, fat_grams, notes} = entry;
+  const {
+    nutrition_id,
+    entry_date,
+    calories_consumed,
+    protein_grams,
+    carbohydrates_grams,
+    fat_grams,
+    notes,
+  } = entry;
   try {
     const sql =
       'UPDATE Nutrition SET entry_date=?, calories_consumed=?, protein_grams=?, carbohydrates_grams=?, fat_grams=?, notes=? WHERE nutrition_id=?';
-    const params = [entry_date, calories_consumed, protein_grams, carbohydrates_grams, fat_grams, notes, nutrition_id];
+    const params = [
+      entry_date,
+      calories_consumed,
+      protein_grams,
+      carbohydrates_grams,
+      fat_grams,
+      notes,
+      nutrition_id,
+    ];
     const [result] = await promisePool.query(sql, params);
     if (result.affectedRows === 0) {
       return {error: 404, message: 'entry not found'};
@@ -101,14 +130,14 @@ const updateNutritionById = async (entry) => {
   }
 };
 
-const updateExerciseById = async (entry) => {
-  const {exercise_id, type, duration, intensity, date} = entry;
+const updateExerciseById = async (exercise_id, userId, entryData) => {
   try {
-    const sql =
-      'UPDATE Exercises SET type=?, duration=?, intensity=?, date=? WHERE exercise_id=?';
-    const params = [type, duration, intensity, date, exercise_id];
+    const params = [entryData, exercise_id, userId];
+    const sql = promisePool.format(
+      `UPDATE Exercises SET ? WHERE exercise_id=? AND user_id=?`,
+      params
+    );
     const [result] = await promisePool.query(sql, params);
-    // console.log(result);
     if (result.affectedRows === 0) {
       return {error: 404, message: 'entry not found'};
     }
@@ -121,13 +150,10 @@ const updateExerciseById = async (entry) => {
   }
 };
 
-
-const updateHrvById = async (entry) => {
-  const {hrv_id, measurement_date, time_of_day, hrv_value, notes} = entry;
+const updateHrvById = async (hrv_id, userId, entryData) => {
   try {
-    const sql =
-      'UPDATE HRVMeasurements SET measurement_date=?, time_of_day=?, hrv_value=?, notes=? WHERE hrv_id=?';
-    const params = [measurement_date, time_of_day, hrv_value, notes, hrv_id];
+    const params = [entryData, hrv_id, userId];
+    const sql = promisePool.format(`UPDATE HRVMeasurements SET ? WHERE hrv_id=? AND user_id=?`, params);
     const [result] = await promisePool.query(sql, params);
     if (result.affectedRows === 0) {
       return {error: 404, message: 'entry not found'};
@@ -141,15 +167,16 @@ const updateHrvById = async (entry) => {
   }
 };
 
-const deleteEntryById = async (id) => {
+const deleteEntryById = async (id, userId) => {
   try {
-    const sql = 'DELETE FROM DiaryEntries WHERE entry_id=?';
-    const params = [id];
+    const sql = 'DELETE FROM DiaryEntries WHERE entry_id=? AND user_id=?';
+    const params = [id, userId];
     const [result] = await promisePool.query(sql, params);
+    // console.log(result);
     if (result.affectedRows === 0) {
-      return {error: 404, message: 'entry not found'};
+      return {error: 404, message: 'Entry not found'};
     }
-    return {message: 'entry deleted', entry_id: id};
+    return {message: 'Entry deleted', entry_id: id};
   } catch (error) {
     console.error('deleteEntryById', error);
     return {error: 500, message: 'db error'};
@@ -171,7 +198,7 @@ const deleteNutritionById = async (id) => {
   }
 };
 
-const deleteMedicationById = async(id) => {
+const deleteMedicationById = async (id) => {
   try {
     const sql = 'DELETE FROM Medications WHERE medication_id=?';
     const params = [id];
@@ -186,7 +213,7 @@ const deleteMedicationById = async(id) => {
   }
 };
 
-const deleteHrvById = async(id) => {
+const deleteHrvById = async (id) => {
   try {
     const sql = 'DELETE FROM HRVMeasurements WHERE hrv_id=?';
     const params = [id];
@@ -201,7 +228,7 @@ const deleteHrvById = async(id) => {
   }
 };
 
-const deleteExerciseById = async(id) => {
+const deleteExerciseById = async (id) => {
   try {
     const sql = 'DELETE FROM Exercises WHERE exercise_id=?';
     const params = [id];
@@ -217,74 +244,90 @@ const deleteExerciseById = async(id) => {
 };
 
 const addExerciseEntry = async (entry) => {
-  const { user_id, date, type, duration, intensity } = entry;
+  const {user_id, date, type, duration, intensity} = entry;
   try {
     const sql = `INSERT INTO Exercises (user_id, date, type, duration, intensity) 
                  VALUES (?, ?, ?, ?, ?)`;
     const params = [user_id, date, type, duration, intensity];
     const [result] = await promisePool.query(sql, params);
     if (result && result.insertId) {
-      return { message: 'Exercise entry added', exercise_id: result.insertId };
+      return {message: 'Exercise entry added', exercise_id: result.insertId};
     } else {
       throw new Error('Failed to add exercise entry');
     }
   } catch (e) {
     console.error('addExerciseEntry error:', e.message);
-    return { error: e.message };
+    return {error: e.message};
   }
 };
 
 const addHrvmeasurement = async (entry) => {
-  const { user_id, measurement_date, time_of_day, hrv_value, notes } = entry;
+  const {user_id, measurement_date, time_of_day, hrv_value, notes} = entry;
   try {
     const sql = `INSERT INTO HRVMeasurements (user_id, measurement_date, time_of_day, hrv_value, notes ) 
                  VALUES (?, ?, ?, ?, ?)`;
-    const params = [user_id, measurement_date, time_of_day, hrv_value, notes ];
+    const params = [user_id, measurement_date, time_of_day, hrv_value, notes];
     const [result] = await promisePool.query(sql, params);
     if (result && result.insertId) {
-      return { message: 'HRV measurement added'};
+      return {message: 'HRV measurement added'};
     } else {
       throw new Error('Failed to add HRV measurement');
     }
   } catch (e) {
     console.error(e.message);
-    return { error: e.message };
+    return {error: e.message};
   }
 };
 
 const postMedication = async (entry) => {
-  const { user_id, name, dosage, frequency, start_date, end_date } = entry;
+  const {user_id, name, dosage, frequency, start_date, end_date} = entry;
   try {
     const sql = `INSERT INTO Medications (user_id, name, dosage, frequency, start_date, end_date ) 
     VALUES (?, ?, ?, ?, ?, ?)`;
-    const params = [user_id, name, dosage, frequency, start_date, end_date ];
+    const params = [user_id, name, dosage, frequency, start_date, end_date];
     const [result] = await promisePool.query(sql, params);
     if (result && result.insertId) {
-      return { message: 'Medications added'};
+      return {message: 'Medications added'};
     } else {
       throw new Error('Failed to add Medications');
     }
   } catch (e) {
     console.error(e.message);
-    return { error: e.message };
+    return {error: e.message};
   }
 };
 
 const postNutrition = async (entry) => {
-  const { user_id, entry_date, calories_consumed, protein_grams, carbohydrates_grams, fat_grams, notes } = entry;
+  const {
+    user_id,
+    entry_date,
+    calories_consumed,
+    protein_grams,
+    carbohydrates_grams,
+    fat_grams,
+    notes,
+  } = entry;
   try {
     const sql = `INSERT INTO Nutrition (user_id, entry_date, calories_consumed, protein_grams, carbohydrates_grams, fat_grams, notes ) 
     VALUES (?, ?, ?, ?, ?, ?, ?)`;
-    const params = [user_id, entry_date, calories_consumed, protein_grams, carbohydrates_grams, fat_grams, notes ];
+    const params = [
+      user_id,
+      entry_date,
+      calories_consumed,
+      protein_grams,
+      carbohydrates_grams,
+      fat_grams,
+      notes,
+    ];
     const [result] = await promisePool.query(sql, params);
     if (result && result.insertId) {
-      return { message: 'Nutrition entry added'};
+      return {message: 'Nutrition entry added'};
     } else {
       throw new Error('Failed to add Nutrition entry');
     }
   } catch (e) {
     console.error(e.message);
-    return { error: e.message };
+    return {error: e.message};
   }
 };
 
@@ -372,6 +415,5 @@ export {
   postNutrition,
   updateNutritionById,
   deleteNutritionById,
-  updateHrvById
-
+  updateHrvById,
 };
